social_media_api/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
Pillow>=10.3
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9 ; platform_system!="Windows"
dj-database-url>=2.1
python-dotenv>=1.0
